third_party/blink/web_tests/vr/getEyeParameters_null.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="../external/wpt/resources/chromium/webxr-test.js"></script>
<script src="resources/vr-test-utils.js "></script>
<script src="resources/test-constants.js"></script>

<canvas id="webgl-canvas"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();
let fakeDisplay = fakeDisplays["FakeMagicWindowOnly"];

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {
    let display = displays[0];

    t.step( () => {
      assert_equals(display.getEyeParameters("left"), null);
      assert_equals(display.getEyeParameters("right"), null);
    }, "Eye parameters are null");

    t.done();
  }, (err) => {
    t.step( () => {
      assert_unreached(err);
    }, "getVRDisplays rejected");
    t.done();
  });
}, [fakeDisplay],
"Test that device which doesn't present has no eye parameters");

</script>
